+++ pyubx2/ubxmessage.py
@@ -47,6 +47,8 @@ class UBXMessage:
         :param object msgClass: message class as str, int or byte
         :param object msgID: message ID as str, int or byte
         :param int msgmode: message mode (0=GET, 1=SET, 2=POLL)
+        :param bool parsebitfield: (kwarg) parse bitfields ('X' type attributes) Y/N
+        :param bool scaling: (kwarg) apply scale factors Y/N
         :param kwargs: optional payload key/value pairs
         :raises: UBXMessageError
 
@@ -60,6 +62,7 @@ class UBXMessage:
         self._checksum = b""
 
         self._parsebf = kwargs.get("parsebitfield", True)  # parsing bitfields Y/N?
+        self._scaling = kwargs.get("scaling", True)  # apply scale factors Y/N?
 
         if msgmode not in (0, 1, 2):
             raise ube.UBXMessageError(f"Invalid msgmode {msgmode} - must be 0, 1 or 2.")
@@ -234,7 +237,7 @@ class UBXMessage:
 
         # if attribute is scaled
         scale = 1
-        if isinstance(att, list):
+        if isinstance(att, list) and self._scaling:
             scale = att[1]
             att = att[0]
 
@@ -249,6 +252,11 @@ class UBXMessage:
             atts = len(self._payload)
         else:
             atts = attsiz(att)
+
+        # if HP element of NAV-HPPOSLLH or NAV-HPPOSECEF message type
+        isNavHP = self._is_navhp(key)
+        if isNavHP:
+            scale = 1
 
         # if payload keyword has been provided,
         # use the appropriate offset of the payload
@@ -271,6 +279,10 @@ class UBXMessage:
 
         setattr(self, keyr, val)
         offset += atts
+
+        # if HP element of NAV-HPPOSLLH or NAV-HPPOSECEF message type
+        if isNavHP:
+            self._set_attribute_navhp(key)
 
         return offset
 
@@ -358,6 +370,69 @@ class UBXMessage:
             setattr(self, keyr, val)
         bfoffset += atts
         return (bitfield, bfoffset)
+
+    def _is_navhp(self, key) -> bool:
+        """
+        Check for NAV-HPPOSLLH or NAV-HPPOSECEF
+        high precision attribute type, e.g.
+        '_latHp' or '_heightHp'.
+
+        :param str key: attribute name e.g. '_latHp'
+        :return: True or False
+        :rtype: bool
+        """
+
+        return (
+            self._ubxClass == b"\x01"
+            and self._ubxID in (b"\x13", b"\x14")
+            and key
+            in (
+                "_lat",
+                "_latHp",
+                "_lon",
+                "_lonHp",
+                "_height",
+                "_heightHp",
+                "_hMSL",
+                "_hMSLHp",
+                "_ecefX",
+                "_ecefXHp",
+                "_ecefY",
+                "_ecefYHp",
+                "_ecefZ",
+                "_ecefZHp",
+            )
+        )
+
+    def _set_attribute_navhp(self, key: str):
+        """
+        Combine separate private standard and high precision
+        attributes of NAV-HPPOSLLH and NAV-HPPOSECEF message
+        types into single public attribute
+        e.g. '_lat' and '_latHp' are combined into 'lat'.
+
+        :param str key: attribute keyword
+        """
+        # pylint: disable=no-member
+
+        if key == "_latHp":
+            val = (self._lat + self._latHp * 0.01) * 1e-7
+        elif key == "_lonHp":
+            val = (self._lon + self._lonHp * 0.01) * 1e-7
+        elif key == "_heightHp":
+            val = self._height + self._heightHp * 0.1  # mm
+        elif key == "_hMSLHp":
+            val = self._hMSL + self._hMSLHp * 0.1  # mm
+        elif key == "_ecefXHp":
+            val = self._ecefX + self._ecefXHp * 0.01  # cm
+        elif key == "_ecefYHp":
+            val = self._ecefY + self._ecefYHp * 0.01  # cm
+        elif key == "_ecefZHp":
+            val = self._ecefZ + self._ecefZHp * 0.01  # cm
+        else:
+            return
+
+        setattr(self, key[1:-2], round(val, ubt.SCALROUND))
 
     def _set_attribute_cfgval(self, offset: int, **kwargs):
         """